Inline normaliseErrors into validateForm
